requirements.txt
aiogram==3.0.0b6
requests==2.28.1
lxml==4.9.2
beautifulsoup4==4.11.1
soupsieve==2.3.2.post1
pip==22.0.2
attrs==22.2.0
pytz==2022.7
idna==3.4
multidict==6.0.4
Babel==2.9.1
pydantic==1.10.2
aiofiles==22.1.0
certifi==2022.12.7
aiohttp==3.8.3
yarl==1.8.2
aiosignal==1.3.1
frozenlist==1.3.3
urllib3==1.26.13
setuptools==59.6.0
psycopg2-binary==2.9.6
schedule~=1.1.0